backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
groq==0.11.0
Pillow==10.1.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
email-validator==2.3.0

# Procesamiento de archivos
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0

